remove redudunt model
