+++ src/main/webapp/app/shared/queries/queries.model.ts
@@ -4,6 +4,22 @@ import { User } from '../user/user.model';
 
 type optionalString = string | null | undefined
 
+export interface QueryGroup {
+    id?: any;
+    name?: string;
+    description?: string;
+    createdDate?: Date;
+    updatedDate?: Date;
+    createdBy?: Subject;
+    updatedBy?: Subject;
+}
+
+export interface QueryParticipant {
+    id?: any;
+    queryGroupId?: number;
+    subjectId?: number;
+    createdBy?: User
+}
 
 export type QueryNode =
   | {
